RessourceGenerator checked
I added 2 variable in the hex class @tokenValue which is the number
corresponding to the hex and has to be compared with the dice number,
and @resourceValue that correspond to the value that the Hex represent.
and I rewrite the assign method to give specifique resource on specic
player on the hex instead of giving the ressource to every player in
listOfPlayer (listOfPlayer contain all player in the hex found with the
same dice number, note that the hex can have a same dice number but
different ressource) that is why I didn't use the listOfPlayer to assign
ressource. instead i check each hex resource in the bank then found
specific player on each vertex if vertex.asset.player is not null then
give him ressource corresponding with the Hex. @makogen  you can try to
compare the class with @CWDavis and figure out how to marge the both
class. they are quite similar. I'm starting now with IA and let you
know.