Player and Cpu
made the isCpu method to verify if the courrent player is a human or the
cpu. also overload the same methode so that you can send a specific
player as param and verify whter that param sent is human or cpu. I also
rewrite the logic of Cpu, the class analyse the resources of the player
but do not take action, instead it set a value in the property action,
that value will be send to the GameScreenController and perform the
action.